% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/birds.R
\docType{data}
\name{birds}
\alias{birds}
\title{Birds}
\format{A data frame with 660 rows and 24 columns:
\describe{
  \item{y}{something}
  \item{I.STATIONARY}{something}
  \item{TIME}{something}
  \item{EFFORT_HRS}{something}
  \item{EFFORT_DISTANCE_KM}{something}
  \item{NUMBER_OBSERVERS}{something}
  \item{ELEV_NED}{something}
  \item{NLCD2006_FS_C11_750_PLAND}{something}
  \item{NLCD2006_FS_C11_750_PLAND}{something}
  \item{NLCD2006_FS_C21_750_PLAND}{something}
  \item{MONTH}{something}
  ...
}}
\source{
Dan Fink of the Cornell Lab of Ornithology \url{http://www.birds.cornell.edu}
}
\usage{
birds
}
\description{
A dataset of birds. 50 observations randomly sampled from each month of 2010, after observations with NAs removed.
}
\keyword{datasets}

