% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppMatch}
\alias{cppMatch}
\title{cppMatch}
\usage{
cppMatch(tr, dist, type)
}
\arguments{
\item{tr}{vector}

\item{dist}{distance}

\item{type}{corresponds to R's representation of distance. 'v' = vector of pairwise distances (class "dist"), 'm' = matrix of pairwise distances (coerced as a vector), 'p' = vector of propensity scores}
}
\description{
Nearest neighbor 1:1 matching within and between conditions.
}
\details{
Ties are broken using the the first match (not randomly).
}

