% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppPredict}
\alias{cppPredict}
\title{cppPredict}
\usage{
cppPredict(x, splitVar, split, lDaughter, rDaughter, nodePred)
}
\arguments{
\item{x}{= matrix of covariates for the test data}

\item{splitVar}{= best split variables}

\item{split}{= best split values}

\item{lDaughter}{= left daughter node assignments}

\item{rDaughter}{= right daughter node assignments}

\item{nodePred}{= terminal node predictions}
}
\description{
Returns predictions from a stored forest.
}
\details{
All other matrices represent the forest. Trees are columns, and nodes are rows.
}

