% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppProx}
\alias{cppProx}
\title{cppProx}
\usage{
cppProx(nodes)
}
\arguments{
\item{nodes}{n x nTree matrix of terminal nodes}
}
\description{
Given an n x nTree matrix of terminal nodes, this function will return an n x n proximimity matrix.
}
\details{
The computed proximities are based on only the out-of-bag predictions. Zero-valued entries in nodes correspond to in-bag instances.
}

