% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastRF.R
\name{fastRF}
\alias{fastRF}
\title{Random forest wrapper function}
\usage{
fastRF(x, y, nTree = 500, replace = TRUE, keepForest = TRUE, mtry = if
  (is.factor(y)) floor(sqrt(ncol(x))) else max(floor(ncol(x)/3), 1),
  nodeSize = if (is.factor(y)) 1 else 5, nSamp = if (replace) nrow(x) else
  ceiling(0.632 * nrow(x)), maxNodes = 2 * trunc(nSamp/max(1, nodeSize - 4)) +
  1)
}
\arguments{
\item{x}{matrix}

\item{y}{vector}

\item{nTree}{number of trees desired, default is 500}

\item{replace}{bootstrap samples or subsamples, default is bootstrap samples}

\item{keepForest}{keep forest or not, default is TRUE}

\item{mtry}{tuning}

\item{nodeSize}{node size}

\item{nSamp}{number of samples}

\item{maxNodes}{maximum number of nodes}
}
\description{
This function is a wrapper for the C++ implementation of random forest
}
\examples{
fastRF()
}
\keyword{causal}
\keyword{forest,}
\keyword{inference}
\keyword{random}

