% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.varIJ.R
\name{forest.varIJ}
\alias{forest.varIJ}
\title{The infinitesimal jackknife for random forests}
\usage{
forest.varIJ(rf)
}
\arguments{
\item{rf}{A random forest trained with replace = TRUE and keep.inbag = TRUE}
}
\description{
Calculate the infinitesimal jackknife variance
}
\examples{
features = matrix(rnorm(100),nrow=10)
response = runif(10)
forestobject = forest(x=features,y=response)
varIJ = forest.varIJ(forestobject)
}
\keyword{forest,}
\keyword{infinitesimal}
\keyword{jackknife}
\keyword{random}
\keyword{variance,}

