% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute.rf.R
\name{impute.rf}
\alias{impute.rf}
\title{This function performs imputation}
\usage{
impute.rf(data, mtry = (length(data) - 2)/2, ...)
}
\arguments{
\item{mtry}{tuning}

\item{x}{matrix}

\item{y}{vector}

\item{nTree}{number of trees desired, default is 500}

\item{replace}{bootstrap samples or subsamples, default is bootstrap samples}

\item{keepForest}{keep forest or not, default is TRUE}

\item{nodeSize}{node size}

\item{nSamp}{number of samples}

\item{maxNodes}{maximum number of nodes}
}
\description{
This function performs imputation
}
\examples{
impute.rf()
}
\keyword{causal}
\keyword{forest,}
\keyword{inference}
\keyword{random}

