% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.rf.R
\name{predict.rf}
\alias{predict.rf}
\title{Prediction wrapper function}
\usage{
\method{predict}{rf}(object, x)
}
\arguments{
\item{object}{random forest output}

\item{x}{matrix}
}
\description{
This function is a wrapper for the C++ implementation of random forest predictions
}
\examples{
predict.rf()
}
\keyword{causal}
\keyword{forest,}
\keyword{inference}
\keyword{random}

