% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictrf.R
\name{predictrf}
\alias{predictrf}
\title{Prediction wrapper function}
\usage{
predictrf(object, x)
}
\arguments{
\item{object}{random forest output}

\item{x}{matrix}
}
\description{
This function is a wrapper for the C++ implementation of random forest predictions
}
\examples{
features = matrix(rnorm(100),nrow=10)
response = runif(10)
rfobject = random.forest(x=features,y=response)
testfeatures = matrix(rnorm(100),nrow=10)
predictrf(object=rfobject,x=testfeatures)
}
\keyword{causal}
\keyword{forest,}
\keyword{inference}
\keyword{random}

