% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.R
\name{proximity}
\alias{proximity}
\title{Proximity matrix function}
\usage{
proximity(terminalNodes, ...)
}
\arguments{
\item{terminalNodes}{n by nTree matrix of terminal nodes}

\item{...}{further arguments passed to or from other methods}
}
\description{
Wrapper for the C++ implementation of random forest proximity matrix. Not ready, do not use.
}
\examples{
n = dim(birds)[1]
n_train = round(n/2)
features = birds[1:n_train,setdiff(names(birds),"detected")]
response = birds[,"detected"]
forestobject = forest(x=features,y=response)
}
\keyword{forest,}
\keyword{matrix}
\keyword{proximity}
\keyword{random}
