% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{proxMatch}
\alias{proxMatch}
\alias{proxMatch-package}
\title{proxMatch}
\description{
Proximity Score Matching using the Random Forest for Causal Inference
}
\section{Description}{

This package implements proximity score matching using a fast implementation of the random forest in C++.
}

\section{Details}{

Functions include: fastRF, rfPredict
}

\section{Dependencies}{

Rcpp and RcppArmadillo
}
\author{
person
}

