% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rf.varIJ.R
\name{rf.varIJ}
\alias{rf.varIJ}
\title{Infinitesimal jacknife variance of random forest}
\usage{
rf.varIJ(object)
}
\arguments{
\item{object}{random forest object}
}
\description{
This function calculates the infinitesimal jacknife variance of the random forest
}
\examples{
rf.varIJ()
}
\keyword{forest,}
\keyword{infitesimal}
\keyword{jacknife}
\keyword{random}
\keyword{variance}

