% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rfPredict.R
\name{rfPredict}
\alias{rfPredict}
\title{Prediction wrapper function}
\usage{
rfPredict(object, x)
}
\arguments{
\item{object}{random forest output}

\item{x}{matrix}
}
\description{
This function is a wrapper for the C++ implementation of random forest predictions
}
\examples{
rfPredict()
}
\keyword{causal}
\keyword{forest,}
\keyword{inference}
\keyword{random}

