% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varInfJacknife.R
\name{varInfJacknife}
\alias{varInfJacknife}
\title{Infinitesimal jacknife variance of random forest}
\usage{
varInfJacknife(object)
}
\arguments{
\item{object}{random forest object}
}
\description{
This function calculates the infinitesimal jacknife variance of the random forest
}
\examples{
varInfJacknife()
}
\keyword{forest,}
\keyword{infitesimal}
\keyword{jacknife}
\keyword{random}
\keyword{variance}

