% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{Random forest wrapper function}
\usage{
forest(x, y, nTree = 500, replace = TRUE, keepForest = TRUE, mtry = if
  (is.factor(y)) floor(sqrt(ncol(x))) else max(floor(ncol(x)/3), 1),
  nodeSize = if (is.factor(y)) 1 else 5, nSamp = if (replace) nrow(x) else
  ceiling(0.632 * nrow(x)), maxNodes = 2 * trunc(nSamp/max(1, nodeSize - 4)) +
  1)
}
\arguments{
\item{x}{matrix}

\item{y}{vector}

\item{nTree}{number of trees desired, default is 500}

\item{replace}{bootstrap samples or subsamples, default is bootstrap samples}

\item{keepForest}{keep forest or not, default is TRUE}

\item{mtry}{tuning}

\item{nodeSize}{node size}

\item{nSamp}{number of samples}

\item{maxNodes}{maximum number of nodes}
}
\description{
Wrapper for the C++ implementation of random forest
}
\examples{
features = birds[,setdiff(names(birds),"y")]
response = birds[,"y"]
forest(x=features,y=response)
}
\keyword{forest}
\keyword{random}

