% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.varU.R
\name{forest.varU}
\alias{forest.varU}
\title{U-statistic based estimate for random forests}
\usage{
forest.varU(object)
}
\arguments{
\item{object}{A random forest trained with replace=FALSE but with common observations}
}
\description{
Calculate the u-statistic based variance
}
\examples{
features = birds[,setdiff(names(birds),"detected")]
response = birds[,"detected"]
forestobject = forest(x=features,y=response,replace=FALSE,var.type="ustat", B=5)
varU = forest.varU(forestobject)
}
\keyword{based}
\keyword{forest,}
\keyword{random}
\keyword{u-statistic}
\keyword{variance,}

