% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.forest.R
\name{predict.forest}
\alias{predict.forest}
\title{Prediction wrapper function}
\usage{
\method{predict}{forest}(object, newdata = NULL, individualTrees = FALSE,
  ...)
}
\arguments{
\item{object}{random forest objectput}

\item{newdata}{matrix}

\item{individualTrees}{whether to return predictions of individual trees, default is FALSE}

\item{...}{further arguments passed to or from other methods}
}
\value{
vector of predictions for each observation if individualTrees=FALSE (default), otherwise matrix of predictions by each tree for each observation
}
\description{
Wrapper for the C++ implementation of random forest predictions
}
\examples{
n = dim(birds)[1]
n_train = round(n/2)
features = birds[1:n_train,setdiff(names(birds),"detected")]
response = birds[,"detected"]
forestobject = forest(x=features,y=response)
testfeatures = birds[n_train+1:n,setdiff(names(birds),"detected")]
predict(object=forestobject,newdata=testfeatures)
}
\references{
Leo Breiman. (2001). Random Forests. Machine Learning 45(1), 5-32. http://link.springer.com/article/10.1023/A:1010933404324
}
\seealso{
\code{\link{forest.varIJ}}, \code{\link{forest.varU}}
}
\author{
Sarah Tan <\email{ht395@cornell.edu}>, David I. Miller
}
\keyword{forest,}
\keyword{prediction}
\keyword{random}
