% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.forest.R
\name{predict.forest}
\alias{predict.forest}
\title{Prediction wrapper function}
\usage{
\method{predict}{forest}(object, newdata = NULL, individualTrees = FALSE,
  ...)
}
\arguments{
\item{object}{random forest objectput}

\item{newdata}{matrix}

\item{individualTrees}{whether to return predictions of individual trees, default is FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Wrapper for the C++ implementation of random forest predictions
}
\examples{
n = dim(birds)[1]
n_train = round(n/2)
features = birds[1:n_train,setdiff(names(birds),"detected")]
response = birds[,"detected"]
forestobject = forest(x=features,y=response)
testfeatures = birds[n_train+1:n,setdiff(names(birds),"detected")]
predict(object=forestobject,newdata=testfeatures)
}
\keyword{forest,}
\keyword{prediction}
\keyword{random}

