% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.forest.R
\name{predict.forest}
\alias{predict.forest}
\title{Prediction wrapper function}
\usage{
\method{predict}{forest}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{random forest output}

\item{newdata}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\description{
Wrapper for the C++ implementation of random forest predictions
}
\examples{
features = matrix(rnorm(100),nrow=10)
response = runif(10)
forestobject = forest(x=features,y=response)
testfeatures = matrix(rnorm(100),nrow=10)
predict(object=forestobject,newdata=testfeatures)
}
\keyword{forest,}
\keyword{prediction}
\keyword{random}

