% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{surfin}
\alias{surfin}
\alias{surfin-package}
\title{surfin}
\description{
Statistical Inference for Random Forests
}
\section{Description}{

This package computes standard error estimates for random forest predictions using a fast implementation of random forests in C++.
}

\section{Details}{

Functions include: rf, predict.rf, impute.rf, proximity.rf, rf.varIJ, rf.varU
}

\section{Dependencies}{

Rcpp and RcppArmadillo
}
\author{
person
}

